feat: add taskPasteIntoPostBox
